requirements.txt
# Core ML
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data
yfinance>=0.2.31
feedparser>=6.0.10

# NLP / Transformers
transformers>=4.35.0
tokenizers>=0.15.0

# Inference Optimization
onnx>=1.15.0
onnxruntime>=1.16.0

# MLOps
mlflow>=2.9.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Technical Indicators
ta>=0.11.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Code Quality
ruff>=0.1.6
mypy>=1.7.0
pre-commit>=3.5.0

# Utilities
pyyaml>=6.0.1
rich>=13.7.0
python-dotenv>=1.0.0
